+++ tests/unit_tests/test_urnc_convert_convert.py
@@ -54,7 +54,9 @@ def test_convert__inDir_solF_forceF_dryF_outExistF_solExistF():
     expected = "tests/expected/minimal-course-converted/lectures"
     urnc.convert.convert(input=input, output=outputs_dir, ask=False)
     cmp = filecmp.dircmp(outputs_dir, expected)
-    assert all((cmp.diff_files == [], cmp.left_only == [], cmp.right_only == []))
+    assert all([cmp.diff_files == [],
+                cmp.left_only == [],
+                all([f.endswith("-solution.ipynb") for f in cmp.right_only])])
 
 
 def test_convert__inNB_solT_forceF_dryF_outExistF_solExistF():
@@ -65,10 +67,8 @@ def test_convert__inNB_solT_forceF_dryF_outExistF_solExistF():
     outputs = [f"{outputs_dir}/urnc.ipynb", f"{outputs_dir}/urnc-solution.ipynb"]
     expects = [f"{expects_dir}/urnc.ipynb", f"{expects_dir}/urnc-solution.ipynb"]
     urnc.convert.convert(input=input, output=outputs_dir, solution=outputs_dir, ask=False)
-    assert all([
-        filecmp.cmp(outputs[0], expects[0]),
-        filecmp.cmp(outputs[1], expects[1])
-    ])
+    assert all([filecmp.cmp(outputs[0], expects[0]),
+                filecmp.cmp(outputs[1], expects[1])])
 
 
 def test_convert__inNB_solT_forceF_dryF_outExistT_solExistT():
@@ -79,10 +79,8 @@ def test_convert__inNB_solT_forceF_dryF_outExistT_solExistT():
     open(outputs[0], "w").write("abc")
     open(outputs[1], "w").write("123")
     urnc.convert.convert(input=input, output=outputs_dir, solution=outputs_dir, ask=False)
-    assert all([
-        open(outputs[0], "r").read() == "abc",
-        open(outputs[1], "r").read() == "123"
-    ])
+    assert all([open(outputs[0], "r").read() == "abc",
+                open(outputs[1], "r").read() == "123"])
 
 
 def test_convert__inNB_solT_forceT_dryF_outExistT_solExistT():